Extract service key converstions to separate utils
