+++ apps/google-analytics-4/package.json
@@ -9,6 +9,7 @@
     "@contentful/react-apps-toolkit": "1.2.10",
     "contentful-management": "10.21.6",
     "emotion": "10.0.27",
+    "lodash": "^4.17.21",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-scripts": "5.0.1"
